Update typed form story to use generic component
